// ModernizacionPersonas.Web/ClientApp/src/app/models/texto-slip.ts
export class TextoSlip {
  codigo: number;
  codigoRamo: number;
  nombreRamo: string;
  codigoSector: number;
  nombreSector: string;
  codigoSubramo: number;
  nombreSubramo: string;
  codigoAmparo: string;
  nombreAmparo: string;
  codigoSeccion: number;
  nombreSeccion: string;
  texto: string;
  esBasico: boolean;
  esAdicional: boolean;
  esEspecial: boolean;
}